--- src/lib.rs
@@ -1,4 +1,0 @@
-pub mod bookmarks;
-pub mod cli;
-pub mod config;
-pub mod finder;
